пошла жара

+++ src/main/java/com/sstu/work/model/User.java
@@ -6,10 +6,11 @@ import lombok.experimental.Accessors;
 @Data
 @Accessors(chain = true)
 public class User {
-    private int Id;
+    private Long Id;
     private String login;
     private String email;
     private String password;
     private Role role;
     private UserInfo info;
+
 }
